Externalize config file reading
it now is done in the runner.
