+++ lib/runner.js
@@ -45,6 +45,10 @@ Runner.prototype.run = function () {
 Runner.prototype._configure = function () {
     debug('configure');
 
+    // import functions
+    var readConfigurationFiles  = require('./runner/readConfigurationFiles');
+    var resolveCapabilities     = require('./runner/resolveCapabilities');
+
     var options = this.options;
 
     if (options.browserstack) {
@@ -58,6 +62,11 @@ Runner.prototype._configure = function () {
         options.wdHost = 'ondemand.saucelabs.com';
         options.wdPort = '80';
     }
+
+    var screenstoryConfig = readConfigurationFiles();
+    options = resolveCapabilities(screenstoryConfig, options);
+
+    this.options = options;
 
     return Promise.resolve(options);
 };
