+++ lib/screenstory.js
@@ -164,7 +164,7 @@ Screenstory.prototype.save = function save(image, options, cb) {
 
     // extract browser id from capabilities
     if (!options.capabilities) {
-        return cb(new VError('Option "capabilities" is required. Just provide '));
+        return cb(new VError('Option "capabilities" is required.\nHINT: You can provide "client.desiredCapabilities".'));
     }
     var id = dasherize(options.id || options.title || this.data.length);
